backend/requirements.txt
# requirements.txt - Versiones más compatibles con Render
# Versiones específicas que funcionan mejor en entornos de producción

# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos MongoDB - VERSIONES MÁS ESTABLES
pymongo==4.3.3
motor==3.1.2
dnspython==2.4.2

# Configuración y validación
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Manejo de archivos y datos
python-multipart==0.0.6

# HTTP requests
httpx==0.25.2

# Validación
email-validator==2.1.0

# Fechas y tiempo
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# Servidor WSGI para producción
gunicorn==21.2.0

# Manejo de archivos async
aiofiles==23.2.1